# Optimized windowed histogram for large window size.