--- a/window.go
+++ b/window.go
@@ -3,8 +3,9 @@ package hdrhistogram
 // A WindowedHistogram combines histograms to provide windowed statistics.
 type WindowedHistogram struct {
 	idx int
-	h   []Histogram
+	h   []*Histogram
 	m   *Histogram
+	tmp *Histogram
 
 	Current *Histogram
 }
@@ -14,12 +15,13 @@ type WindowedHistogram struct {
 func NewWindowed(n int, minValue, maxValue int64, sigfigs int) *WindowedHistogram {
 	w := WindowedHistogram{
 		idx: -1,
-		h:   make([]Histogram, n),
+		h:   make([]*Histogram, n),
 		m:   New(minValue, maxValue, sigfigs),
+		tmp: New(minValue, maxValue, sigfigs),
 	}
 
 	for i := range w.h {
-		w.h[i] = *New(minValue, maxValue, sigfigs)
+		w.h[i] = New(minValue, maxValue, sigfigs)
 	}
 	w.Rotate()
 
@@ -29,17 +31,21 @@ func NewWindowed(n int, minValue, maxValue int64, sigfigs int) *WindowedHistogra
 // Merge returns a histogram which includes the recorded values from all the
 // sections of the window.
 func (w *WindowedHistogram) Merge() *Histogram {
-	w.m.Reset()
-	for _, h := range w.h {
-		w.m.Merge(&h)
-	}
-	return w.m
+	w.tmp.Reset()
+	w.tmp.Merge(w.m)
+	w.tmp.Merge(w.Current)
+	return w.tmp
 }
 
 // Rotate resets the oldest histogram and rotates it to be used as the current
-// histogram.
-func (w *WindowedHistogram) Rotate() {
+// histogram. Returns merged histogram the same Merge would return.
+func (w *WindowedHistogram) Rotate() *Histogram {
+	if w.Current != nil {
+		w.m.Merge(w.Current)
+	}
 	w.idx++
-	w.Current = &w.h[w.idx%len(w.h)]
+	w.Current = w.h[w.idx%len(w.h)]
+	w.m.Unmerge(w.Current)
 	w.Current.Reset()
+	return w.m
 }